+++ psTerminalPerfCounter/Classes/ServerConfiguration.ps1
@@ -1,0 +1,10 @@
+using namespace System.Collections.Generic
+
+class ServerConfiguration {
+    [string]                $serverName
+
+    ServerConfiguration([string]$serverName) {
+        $this.serverName     = $serverName
+    }
+
+}